Update swagger API documentation and delete car enpoint.

+++ src/main/java/com/unicorn/rental/service/car/CarService.java
@@ -53,6 +53,12 @@ public class CarService {
     }
 
     public void deleteCarById(int id) {
+
+        Optional<Car> car = carDao.findById(id);
+        if (car.isEmpty()) {
+            throw new ItemNotFoundException();
+        }
+
         carDao.deleteById(id);
     }
